Add IdoMenuItemFactory
It's implementing the GtkMenuItemFactory extension point (only available in
Ubuntu's version of gtk+). It doesn't create any menu items yet.

The extension point must be registered before calling gtk_menu_new_from_model()
with a menu model that references any of the custom menu items ido will
provide. Registering means the type must exist in the type system. This patch
adds an ido_init() which does that. Consumers of libido are required to call
this function from now on.