Add IdoActionHelper
This class contains some functionality that all menu items that are bound to
an action need. It listens for action additions and removals, changes in
enabled state, and state changes. It is _not_ intended to be used by specific
menu items, but rather as the glue in IdoMenuItemFactory, to bind a widget to
an action.

An alternative way to reach the same goal is to have an IdoMenuItem base class
which contains the common functionality. This would have required touching
all existing menu items (slider, user menu item, media player items) and
would have added more complexity to them.

It is similar in spirit to GtkActionHelper (a private class in gtk), but very
different in API and implementation.