idoscalemenuitem: don't update the action state when the action state changes
GtkRange has the same weird semantics for its "value-changed" signal that the
check menu items have: it is emitted not only on user interaction, but also
when gtk_range_set_value() is called.

The handler to "value-changed" updates the action. gtk_range_set_value() is
called when the action notifies about a state change. Loop, meet loop.

This patch works around that by adding a "value-changed" to IdoScaleMenuItem
that is only emitted when the user changes the slider's value.