IdoUserMenuItem: only allow file icons as avatars
In practice, all avatars are file icons. This restriction allowed for quite
some cleanup, as we don't need to worry about symbolic icons and themes aymore.

Note: this patch removes the warning when the user avatar image was not found.
This is a common enough case that the warning just clutters up logs.